+++ visvis2/materials/_mesh.py
